requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
black==25.1.0
click==8.1.8
discord==2.3.2
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
isort==6.0.1
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.3.0
python-dotenv==1.0.1
yarl==1.18.3
yt-dlp==2025.2.19
google-generativeai==0.5.2